move methods to progress helper
