+++ default_settings.php
@@ -25,17 +25,14 @@
  */
 use plagiarism_unplag\classes\unplag_notification;
 
-if (!defined('MOODLE_INTERNAL')) {
-    die('Direct access to this script is forbidden.');
-}
+require_once(dirname(dirname(__FILE__)) . '/../config.php');
+require_once(dirname(__FILE__) . '/lib.php');
+require_once(dirname(__FILE__) . '/unplag_form.php');
 
 global $CFG, $DB, $OUTPUT;
 
-require_once(dirname(dirname(__FILE__)) . '/../config.php');
 require_once($CFG->libdir . '/adminlib.php');
 require_once($CFG->libdir . '/plagiarismlib.php');
-require_once(dirname(__FILE__) . '/lib.php');
-require_once(dirname(__FILE__) . '/unplag_form.php');
 
 require_login();
 admin_externalpage_setup('plagiarismunplag');
